+++ public/admin/tool/muprog/lib.php
@@ -153,7 +153,7 @@ function tool_muprog_get_fontawesome_icon_map() {
         'tool_muprog:itemcourse' => 'fa-graduation-cap',
         'tool_muprog:itemset' => 'fa-list',
         'tool_muprog:itemtop' => 'fa-cubes',
-        'tool_muprog:itemtraining' => 'fa-ellipsis-h',
+        'tool_muprog:itemtraining' => 'fa-solid fa-grip',
         'tool_muprog:move' => 'fa-arrows',
         'tool_muprog:program' => 'fa-cubes',
         'tool_muprog:myprograms' => 'fa-cubes',